requirements.txt
python-dotenv
requests
crewai
langchain
openai
twilio
pymongo
flask 